update(lucifer-*): validate user update with yup on both sides
