+++ apps/lucifer-api/src/users/users.controller.ts
@@ -1,14 +1,14 @@
-import { Body, Controller, Get, Put, UseFilters, UseGuards, ValidationPipe } from '@nestjs/common';
+import { Body, Controller, Get, Put, UseFilters, UseGuards } from '@nestjs/common';
 import { AuthGuard } from '@nestjs/passport';
 
+import { IUpdateUser, updateUserSchema, User } from '@lucifer/types';
 import { AllowIf, ScopeGuard, Scopes } from '../auth/scope.guard';
+import { Context, Ctx } from '../context';
+import { YupPipe } from '../utils/yup.pipe';
 
-import { User } from '@lucifer/types';
 import { Auth0ErrorFilter } from './auth0-error.filter';
-import { UpdateUser } from './user.schema';
 import { UsersService } from './users.service';
 import { UserId } from './user-id.param';
-import { Context, Ctx } from '../context';
 
 // Controller
 @Controller('/users')
@@ -40,7 +40,7 @@ export class UsersController {
   async update(
     @Ctx() ctx: Context,
     @UserId('id') id: string,
-    @Body(ValidationPipe) update: UpdateUser
+    @Body(new YupPipe(updateUserSchema)) update: IUpdateUser
   ): Promise<User> {
     return await this.users.update(ctx, id, update);
   }
