+++ libs/types/src/index.ts
@@ -1,4 +1,5 @@
 export * from './lib/machine';
 export * from './lib/project';
-export * from './lib/user';
-export * from './lib/variable';
+export * from './permissions';
+export * from './user';
+export * from './variable';
