+++ libs/types/src/variable/entity.ts
@@ -1,0 +1,11 @@
+// Interface
+export interface IVariable {
+  // Id
+  id: string;
+  projectId: string;
+  adminId: string;
+
+  // Data
+  name: string;
+  value: string;
+}
